refactor(web): extract the clipboard family into useClipboardActions
PR 3 of the SpatialEditor decomposition (3703 -> 3563 lines). Moves
copy/cut/paste/duplicate plus the foreign-text fallback note into
use-clipboard-actions.ts, behind the one seam the plan called for:
selectNodes(ids) — set-members through the selection reducer plus
dropping any edge selection — so the hook never writes selection state
directly and the primary/extras invariant stays enforced in the editor.
The PendingCut shape moves with the cut it describes.

useEditorMeasurements moves up beside the viewport declaration so
viewportCenterScreen is initialized before the hook call reads it
(same call, earlier line — hooks stay unconditional).

purity-guard.test.ts's command-boundary tripwire now walks the editor
PLUS every hook its onChange call sites moved to; mutation-checked — a
hand-rolled canvas spread reaching onChange from the hook fails the
guard naming the file.

Behavior-preserving: spatial-editor jsdom 42 files / 470 tests, and the
clipboard-covering browser suites (clipboard, os-clipboard,
SpatialEditor.browser, context-menu - 81 tests) pass un-weakened.
